+++ src/main/java/com/kentington/thaumichorizons/client/renderer/tile/TilePortalTHRender.java
@@ -76,40 +76,24 @@ public class TilePortalTHRender extends TileEntitySpecialRenderer {
             final double px = x + 0.5;
             final double py = y + 0.5;
             final double pz = z + 0.5;
-            final Vec3 v1 = Vec3.createVectorHelper((double) (-arX - arYZ), (double) (-arXZ), (double) (-arZ - arXY));
-            final Vec3 v2 = Vec3.createVectorHelper((double) (-arX + arYZ), (double) arXZ, (double) (-arZ + arXY));
-            final Vec3 v3 = Vec3.createVectorHelper((double) (arX + arYZ), (double) arXZ, (double) (arZ + arXY));
-            final Vec3 v4 = Vec3.createVectorHelper((double) (arX - arYZ), (double) (-arXZ), (double) (arZ - arXY));
+            final Vec3 v1 = Vec3.createVectorHelper(-arX - arYZ, -arXZ, -arZ - arXY);
+            final Vec3 v2 = Vec3.createVectorHelper(-arX + arYZ, arXZ, -arZ + arXY);
+            final Vec3 v3 = Vec3.createVectorHelper(arX + arYZ, arXZ, arZ + arXY);
+            final Vec3 v4 = Vec3.createVectorHelper(arX - arYZ, -arXZ, arZ - arXY);
             final int frame = (int) time % 16;
             final float f2 = frame / 16.0f;
             final float f3 = f2 + 0.0625f;
             final float f4 = 0.0f;
             final float f5 = 1.0f;
             tessellator.setNormal(0.0f, 0.0f, -1.0f);
-            tessellator.addVertexWithUV(
-                    px + v1.xCoord * scale,
-                    py + v1.yCoord * scaley,
-                    pz + v1.zCoord * scale,
-                    (double) f2,
-                    (double) f5);
-            tessellator.addVertexWithUV(
-                    px + v2.xCoord * scale,
-                    py + v2.yCoord * scaley,
-                    pz + v2.zCoord * scale,
-                    (double) f3,
-                    (double) f5);
-            tessellator.addVertexWithUV(
-                    px + v3.xCoord * scale,
-                    py + v3.yCoord * scaley,
-                    pz + v3.zCoord * scale,
-                    (double) f3,
-                    (double) f4);
-            tessellator.addVertexWithUV(
-                    px + v4.xCoord * scale,
-                    py + v4.yCoord * scaley,
-                    pz + v4.zCoord * scale,
-                    (double) f2,
-                    (double) f4);
+            tessellator
+                    .addVertexWithUV(px + v1.xCoord * scale, py + v1.yCoord * scaley, pz + v1.zCoord * scale, f2, f5);
+            tessellator
+                    .addVertexWithUV(px + v2.xCoord * scale, py + v2.yCoord * scaley, pz + v2.zCoord * scale, f3, f5);
+            tessellator
+                    .addVertexWithUV(px + v3.xCoord * scale, py + v3.yCoord * scaley, pz + v3.zCoord * scale, f3, f4);
+            tessellator
+                    .addVertexWithUV(px + v4.xCoord * scale, py + v4.yCoord * scaley, pz + v4.zCoord * scale, f2, f4);
             tessellator.draw();
         }
         GL11.glDisable(3042);
