+++ src/main/java/com/kentington/thaumichorizons/common/lib/GatewayTeleporter.java
@@ -27,6 +27,6 @@ public class GatewayTeleporter extends Teleporter {
 
     public void placeInPortal(final Entity p_77185_1_, final double p_77185_2_, final double p_77185_4_,
             final double p_77185_6_, final float p_77185_8_) {
-        p_77185_1_.setPositionAndRotation(this.x + 0.5, (double) this.y, this.z + 0.5, this.yaw, 0.0f);
+        p_77185_1_.setPositionAndRotation(this.x + 0.5, this.y, this.z + 0.5, this.yaw, 0.0f);
     }
 }
